requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
mysql-connector-python==9.0.0
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.1
wheel==0.44.0
